Improve formulation on interface type conversion
Since `.(TYPE)` merely extracts the explicit type of the interface, try
to indicate more clearly that this is not a typecast.

This could be written even more detailed, but it seems like a good idea
to avoid going too deep. If reader needs more info, Effective Go[1]
provides a good explanation.

[1] https://golang.org/doc/effective_go.html#interface_conversions