Use concurrency example that demonstrates problem
I found that I couldn't get it to produce output other than 1n2 when I ran it.
This code produces erratic output a lot of the time; which I feel better shows
the problems of bad concurrency. Someone less familiar with concurrency could
be led to believe that these effects are unlikely enough to be ignored with
the current example I feel.